+++ sdk/rollup.test.js
@@ -39,6 +39,11 @@ export default networks.map((network) => {
             // Used by the SDK
             "comlink",
             `@provablehq/wasm/${network}.js`,
+            "core-js/proposals/json-parse-with-source.js",
+
+            // Used by tests
+            "chai",
+            "sinon",
         ],
         plugins: [
             replace({
