+++ sdk/tests/record-provider.test.ts
@@ -1,7 +1,6 @@
 import { expect } from "chai";
 import {Account, AleoNetworkClient, BlockHeightSearch, NetworkRecordProvider} from "../src/node";
 import {beaconPrivateKeyString} from "./data/account-data";
-import {log} from "console";
 
 describe.skip('RecordProvider', () => {
     let account: Account;
@@ -17,13 +16,8 @@ describe.skip('RecordProvider', () => {
     describe('Record provider', () => {
         it('should not find records where there are none', async () => {
             const params = new BlockHeightSearch(0, 100);
-            try {
-                const records = await recordProvider.findCreditsRecords([100, 200], true, [], params);
-                expect(<object>records).equal([])
-            } catch (e) {
-                log(e)
-                throw e;
-            }
+            const records = await recordProvider.findCreditsRecords([100, 200], true, [], params);
+            expect(<object>records).equal([]);
         });
     });
 });
